added Navbar not finished yet
