+++ src/App.tsx
@@ -4,6 +4,7 @@ import Home from "./pages/Home/Home";
 import MovieInfo from "./pages/MovieInfo/MovieInfo";
 import Booking from "./pages/Booking/Booking";
 import Admin from "./pages/Admin/Admin";
+import Login from "./pages//Login/Login";
 import {BrowserRouter, Routes,Route} from 'react-router-dom';
 
 function App() {
@@ -14,6 +15,7 @@ function App() {
               <Route path="MovieInfo" element={<MovieInfo/>}/>
               <Route path="Booking" element={<Booking/>}/>
               <Route path="Admin" element={<Admin/>}/>
+              <Route path="Login" element={<Login/>}/>
           </Routes>
       </BrowserRouter>
