+++ src/components/Navbar/Navbar.css
@@ -1,39 +1,28 @@
-
-.navBar{
-    width: 100%;
-    background: #293133;
-}
-
-.search {
-    width: 50%;
-    margin-top: 1%;
-    margin-left:1%;
-    margin-bottom: 1%;
+.navbar{
     display: flex;
+    justify-content: space-around;
     align-items: center;
-    justify-content: center;
-
-    padding: 1.5rem 1.75rem;
-    border-radius: 3rem;
-    background: darkgrey;
-    -webkit-box-shadow: 5px 5px 7px #1c1d1f, -5px -5px 7px #222527;
-    box-shadow: 5px 5px 7px #1c1d1f, -5px -5px 7px #222527;
+    min-height: 10vh;
+    background: #212426;
+    color: grey;
 }
-
-.search input {
-    flex: 1;
-    border: none;
+.links{
+    width: 70%;
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    list-style: none;
+}
+.linkText{
+    color: darkgrey;
+    text-decoration: none;
     font-size: 1.5rem;
     font-family: var(--font-raleway);
     font-weight: 500;
-    padding-right: 1rem;
-    outline: none;
-    color: black;
-    background: darkgrey;
 }
-
-.search img {
-    width: 35px;
-    height: 35px;
-    cursor: pointer;
+.linkText li:hover{
+    color: #61dafb;
+}
+.linkText li:active{
+    color: #61dafb;
 }
