+++ src/components/Searchbar/Searchbar.css
@@ -1,0 +1,36 @@
+
+
+
+.searchbar {
+    width: 100%;
+    margin-top: 1%;
+    margin-left:1%;
+    margin-bottom: 1%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    padding: 1.5rem 1.75rem;
+    border-radius: 3rem;
+    background: darkgrey;
+    -webkit-box-shadow: 5px 5px 7px #1c1d1f, -5px -5px 7px #222527;
+    box-shadow: 5px 5px 7px #1c1d1f, -5px -5px 7px #222527;
+}
+
+.searchbar input {
+    flex: 1;
+    border: none;
+    font-size: 1.5rem;
+    font-family: var(--font-raleway);
+    font-weight: 500;
+    padding-right: 1rem;
+    outline: none;
+    color: black;
+    background: darkgrey;
+}
+
+.searchbar img {
+    width: 35px;
+    height: 35px;
+    cursor: pointer;
+}
