+++ src/pages/Admin/Admin.tsx
@@ -1,8 +1,9 @@
 import "./Admin.css"
-
+import Navbar from "../../components/Navbar/Navbar";
 function Admin(){
     return(
       <div className="app">
+          <Navbar/>
           <p>Hier kommt die Admin Seite rein </p>
       </div>
     );
