+++ src/pages/Booking/Booking.tsx
@@ -1,8 +1,11 @@
 import './Booking.css'
+import Navbar from "../../components/Navbar/Navbar";
 
 function Booking(){
     return(
+
       <div className="app">
+          <Navbar/>
           <p>
               hier Sitzplätze buchen
           </p>
