+++ src/pages/Home/Home.css
@@ -17,14 +17,7 @@ body {
     background-color: #212426;
 }
 
-.app {
-    padding: 0;
 
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-}
 
 @media screen and (max-width: 600px) {
     .app {
