+++ src/pages/Home/Home.tsx
@@ -1,12 +1,13 @@
-import Navbar from "../../components/Navbar/Navbar";
+
 import Movie from "../../components/Movie/Movie";
 import "./Home.css"
+import Navbar from "../../components/Navbar/Navbar";
 function Home(){
 
 
     return(
         <div className="app">
-               <Navbar/>
+            <Navbar/>
                 <h1>Veranstaltungen</h1>
                 <Movie/>
                 <Movie/>
