+++ src/pages/Login/Login.tsx
@@ -1,0 +1,11 @@
+import "./Login.css"
+import Navbar from "../../components/Navbar/Navbar";
+function Admin(){
+    return(
+        <div className="app">
+            <Navbar/>
+            <p>Login </p>
+        </div>
+    );
+}
+export default Admin
